Testing add to cart

+++ backend/custom_modules/user-transac.js
@@ -258,29 +258,45 @@ async function getCart(info, callback) {
         })
         .catch(err => {
             callback(err, null);
-        })
+        });
 }
 
 async function addToCart(info, callback) {
-    var query = "insert into shoppingcart values(?,?,?,?,?,?)";
-    var fields = [info.userid, info.bookid, info.quantity, info.price, info.total, info.title];
+    var query = 'insert into shoppingcart values(?,?,?,?,?,?)';
+    var fields = [
+        info.userid,
+        info.bookid,
+        info.quantity,
+        info.price,
+        info.total,
+        info.title
+    ];
 
-    pool.query(query, fields).then(res => {
-        callback(null,2); //book added
-    }).catch(err => {
-        if (err.errno === NOT_UNIQUE) {
-            callback(null, 1) //book already in the cart
-        }
-        else {
-            callback (err,null);
-        }
-    })
+    pool.query(query, fields)
+        .then(res => {
+            callback(null, 2); //book added
+        })
+        .catch(err => {
+            if (err.errno === NOT_UNIQUE) {
+                callback(null, 1); //book already in the cart
+            } else {
+                callback(err, null);
+            }
+        });
 }
 
 async function editQuantity(info, callback) {
-    var query = "update shoppingcart set quantity=?, price=?, total=?, title=? where userid=? and bookid=?";
+    var query =
+        'update shoppingcart set quantity=?, price=?, total=?, title=? where userid=? and bookid=?';
 
-    var fields = [info.quantity, info.price, info.total, info.title, info.userid, info.bookid];
+    var fields = [
+        info.quantity,
+        info.price,
+        info.total,
+        info.title,
+        info.userid,
+        info.bookid
+    ];
     pool.query(query, fields)
         .then(res => {
             callback(null, res);
@@ -291,21 +307,18 @@ async function editQuantity(info, callback) {
 }
 
 async function delCartItems(info, callback) {
-    var query =
-        "delete from shoppingcart where (" +
-        "userid=? and bookid=?)";
+    var query = 'delete from shoppingcart where (' + 'userid=? and bookid=?)';
 
     var data = [info.userid, info.bookid];
-    console.log("I WORKED");
+    console.log('I WORKED');
     pool.query(query, data)
         .then(res => {
             callback(null); // query successful
         })
         .catch(err => {
             console.log(err);
-        })
+        });
 }
-
 
 module.exports = {
     createUser,
@@ -321,5 +334,5 @@ module.exports = {
     getCart,
     addToCart,
     delCartItems,
-    editQuantity,
+    editQuantity
 };
